requirements.txt
click>=5.1
pyvcf>=0.6.7
pip
requests